Add a way to completely disable the default placeholder plugin
I'd like to use slate-react, along with a custom placeholder.
The problem I'm running into though is that slate-react _always_ adds the PlaceholderPlugin pretty much.  Even setting the placeholder to an empty string will still actually add a magic div to the dom, and that's what breaking for me.

So, I figured adding support for passing _exactly_ null in to disable the plugin would work well.  Folks that don't pass in anything will still get the "default" of an "empty placeholder" so hopefully nothing breaks.