Add better support for complex imput models like composition/ime/auto-correct/etc
Uses a similar model to editors like etherpad, which let the browser modify the dom, and then incorporates any changes in the browser into the editor's internal data structure.

The before.js plugin has most of the documentation, so I'd probably start there when reading. I'm also happy to sit down with you in person to go through this one.